+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix splitting media types with quoted commas
+
 0.5.1 / 2015-02-14
 ==================